refactor : modularizing broadcasting
