+++ lib/receptor.js
