+++ lib/sockets.js
