+++ package.json
@@ -22,11 +22,11 @@
   },
   "homepage": "https://github.com/ygarasab/packet-rebroadcaster#readme",
   "dependencies": {
+    "buffer-broadcaster": "^1.0.0",
     "pcap": "^3.1.0",
-    "ws": "^7.3.1",
     "yargs": "^15.4.1"
   },
   "bin": {
-    "rebroadcast-packets": "./bin/rebroadcast-packets.js"
+    "rebroadcast-packets": "./bin/index.js"
   }
 }
